+++ www/public/mysql.php
@@ -1,0 +1,14 @@
+<?php
+$servername = "192.168.99.100";
+$username = "root";
+$password = "root";
+
+try {
+    $conn = new PDO("mysql:host=$servername", $username, $password);
+    // set the PDO error mode to exception
+    $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+    echo "Connected successfully";
+} catch(PDOException $e) {
+    echo "Connection failed: " . $e->getMessage();
+}
+?>