+++ CHANGELOG.md
@@ -2,4 +2,5 @@ Changelog
 =========
  
 ## 0.0.*
+ * Add Docker configure for Nginx + PHP + MySQL
  * Add bash init script and folder structure